+++ src/pages/index.module.css
@@ -1,0 +1,18 @@
+.container {
+    max-width: 800px;
+    margin: 0 auto;
+}
+
+.h1 {
+    font-weight: normal;
+    text-align: center;
+    font-size: 1.5rem;
+    margin: 10px;
+}
+
+.tile {
+    background-color: #dddddd;
+    height: 400px;
+    margin: 50px 10px;
+    border-radius: 10px;
+}
